requirements.txt
google-genai==1.7.0
google-cloud-aiplatform
pillow
python-dotenv
streamlit
requests
pydantic